msm: kgsl: Change local vars being assigned from size_t to size_t
A size_t variable being assigned to an integer variable results in
truncation which is problematic. Fix this by changing all those to
size_t.

+++ drivers/gpu/msm/kgsl_iommu.c
@@ -1756,7 +1756,7 @@ kgsl_iommu_map(struct kgsl_pagetable *pt,
 	int ret;
 	unsigned int iommu_virt_addr;
 	struct kgsl_iommu_pt *iommu_pt = pt->priv;
-	int size = memdesc->size;
+	size_t size = memdesc->size;
 
 	BUG_ON(NULL == iommu_pt);
 
@@ -1765,7 +1765,7 @@ kgsl_iommu_map(struct kgsl_pagetable *pt,
 	ret = iommu_map_range(iommu_pt->domain, iommu_virt_addr, memdesc->sg,
 				size, protflags);
 	if (ret) {
-		KGSL_CORE_ERR("iommu_map_range(%p, %x, %p, %d, %x) err: %d\n",
+		KGSL_CORE_ERR("iommu_map_range(%p, %x, %p, %zd, %x) err: %d\n",
 			iommu_pt->domain, iommu_virt_addr, memdesc->sg, size,
 			protflags, ret);
 		return ret;
